requirements.txt
fastapi
uvicorn
anyio
crawl4ai  # Ensure the version is correct, use pip freeze to get it if needed
jsonschema  # If not already installed, used by crawl4ai for JSON schema
requests # If crawl4ai makes HTTP requests internally (check dependencies)
selenium # If using the selenium webdriver (depends on crawl4ai config)
chromium # Ensure compatibility for chromium browser used in crawl4ai
python-dotenv==1.0.0
google-generativeai>=0.3.0